requirements.txt
Django==5.1.3
django-simple-captcha
pillow
